Add Sacha Chua's weekly review support.
I have massaged the available code while trying to understand it. There
are small changes to the regexps used to pick up last and next week's
TODO items. Functionally speaking, I tried to be faithful to the
original functionality of the original entry points - but there are no
guarantees.

There is also a small agenda configuration change.